+++ src/components/shared/Combobox/Combobox.tsx
@@ -57,6 +57,7 @@ function LunaticComboBox({
 	onSelect,
 	value,
 	options,
+	shouldDisplayOptionsId = true,
 	messageError,
 	search = EMPTY_SEARCH,
 	getOptionValue = getDefaultOptionValue,
@@ -184,6 +185,7 @@ function LunaticComboBox({
 					isLoading={isLoading}
 					optionRenderer={optionRenderer}
 					options={options}
+					shouldDisplayOptionsId={shouldDisplayOptionsId}
 					focused={focused}
 					selectedIndex={selectedIndex}
 					expanded={expanded}
@@ -212,15 +214,14 @@ function getIndexFromOptions(
 	if (!Array.isArray(options)) {
 		return undefined;
 	}
-	return options.map(getOptionValue).findIndex((v) => v === value);
+	return options.map(getOptionValue).indexOf(value ?? '');
 }
 
 /**
  * Extract value from an option item
  */
-function getDefaultOptionValue(option: ComboboxOptionType = { value: '' }) {
-	const { id, value } = option;
-	return id || value;
+function getDefaultOptionValue(option?: ComboboxOptionType) {
+	return option?.id || option?.value || '';
 }
 
 export const Combobox = slottableComponent('Combobox', LunaticComboBox);