+++ src/components/shared/Combobox/Panel/ComboboxOption.spec.tsx
@@ -14,13 +14,35 @@ describe('DefaultOptionRenderer', () => {
 		expect(labelElement).toBeNull();
 	});
 
-	it('renders with label', () => {
+	it('renders with label and displays id by default', () => {
 		const option = { id: '1', value: 'Value', label: 'Label' };
 		const { getByText } = render(<ComboboxOption option={option} />);
-		const idElement = getByText(option.id);
-		const labelElement = getByText(option.label);
-		expect(idElement).toBeInTheDocument();
-		expect(labelElement).toBeInTheDocument();
+
+		expect(getByText(option.id)).toBeInTheDocument();
+		expect(getByText('-')).toBeInTheDocument();
+		expect(getByText(option.label)).toBeInTheDocument();
+	});
+
+	it('renders with label and displays id when prop is true', () => {
+		const option = { id: '1', value: 'Value', label: 'Label' };
+		const { getByText } = render(
+			<ComboboxOption option={option} shouldDisplayOptionId={true} />
+		);
+
+		expect(getByText(option.id)).toBeInTheDocument();
+		expect(getByText('-')).toBeInTheDocument();
+		expect(getByText(option.label)).toBeInTheDocument();
+	});
+
+	it('renders with label and does not display id when prop is false', () => {
+		const option = { id: '1', value: 'Value', label: 'Label' };
+		const { getByText, queryByText } = render(
+			<ComboboxOption option={option} shouldDisplayOptionId={false} />
+		);
+
+		expect(queryByText(option.id)).not.toBeInTheDocument();
+		expect(queryByText('-')).not.toBeInTheDocument();
+		expect(getByText(option.label)).toBeInTheDocument();
 	});
 
 	it('renders with selected class', () => {